// client/src/models/TeamMemberModel.tsx
export interface TeamMember {
  name: string;
  title: string;
  imageUrl: string;
  bio?: string; // Optional bio property
  email?: string; // Optional email property
  github?: string; // Optional GitHub profile URL
  linkedin: string; // Optional LinkedIn profile URL
  instagram?: string; // Optional Instagram profile URL
  twitter?: string; // Optional Twitter profile URL
}